+++ ksp/sealed/generator/test/src/commonMain/kotlin/TestSealedTypesWorkaround.kt
@@ -1,0 +1,12 @@
+// THIS CODE HAVE BEEN GENERATED AUTOMATICALLY
+// TO REGENERATE IT JUST DELETE FILE
+// ORIGINAL FILE: Test.kt
+package dev.inmo.micro_utils.ksp.`sealed`.generator.test
+
+import kotlin.collections.Set
+import kotlin.reflect.KClass
+
+private val subtypes: Set<KClass<out Test>> = setOf(Test.C::class,
+    Test.A::class)
+
+public fun Test.Companion.subtypes(): Set<KClass<out Test>> = subtypes